+++ lib/vector_tile.dart
@@ -1,6 +1,16 @@
+import 'dart:math';
+
 import 'package:meta/meta.dart';
-import 'package:vector_tile/raw/raw_vector_tile.dart' as raw;
 import 'package:vector_tile/vector_tile_layer.dart';
+import 'package:vector_tile/raw/raw_vector_tile.dart' as raw;
+import 'package:vector_tile/util/geojson.dart';
+
+export 'package:vector_tile/vector_tile_geom_type.dart';
+export 'package:vector_tile/vector_tile_value.dart';
+export 'package:vector_tile/vector_tile_feature.dart';
+export 'package:vector_tile/vector_tile_layer.dart';
+export 'package:vector_tile/util/geojson.dart';
+export 'package:vector_tile/util/geometry.dart';
 
 class VectorTile {
   raw.VectorTile rawTile;
@@ -18,5 +28,22 @@ class VectorTile {
     }).toList();
 
     return VectorTile(rawTile: rawTile, layers: layers);
+  }
+
+  GeoJsonFeatureCollection toGeoJson({@required int x, @required int y, @required int z}) {
+    List<GeoJson> featuresGeoJson = [];
+    this.layers.forEach((layer) {
+      int size = layer.extent * pow(2, z);
+      int x0 = layer.extent * x;
+      int y0 = layer.extent * y;
+
+      layer.features.forEach((feature) {
+        featuresGeoJson.add(
+          feature.toGeoJsonWithExtentCalculated(x0: x0, y0: y0, size: size)
+        );
+      });
+    });
+
+    return GeoJsonFeatureCollection(features: featuresGeoJson);
   }
 }